+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Changed
+
+- Allow experiment name to be `None` when creating new experiments via `Beaker.experiment.create()`.
+
 ## [v1.21.0](https://github.com/allenai/beaker-py/releases/tag/v1.21.0) - 2023-09-08
 
 ### Added